+++ tmx/lib/skylab/tmx/reports-/punchlist.rb
@@ -1,0 +1,10 @@
+module Skylab::TMX
+
+  class Reports_::Punchlist
+
+    def inititialize(*)
+      self._JUST_A_STUB
+    end
+  end
+end
+# #history: born to replace #tombstone: the static punchlist.template